iso8601: FixedOffset objects can not be unpickled
This patch replaces the __init__ of FixedOffset with a set method
instead. This is so we can unpickle datetime objects using this tzinfo
based class as a tzinfo= argument. you can read all about it at
https://docs.python.org/2/library/datetime.html#tzinfo-objects

I ran into this while trying to unpickle a pickled dictionary filled with
UrlSetElement objects that I had read in from the sitemap library which
had a lastmod set in the parsed sitemap you get a nasty traceback
complaining about __init__ requiring 4 arguments but only one given with
out much useful info on what inside the pickle file caused it.